Moved UnevaluatedConstData to chalk-integration
